feat: Move internal_cpp and internal_python under internal
